requirements.txt
# AI Football Betting Advisor - Required Packages
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-telegram-bot==20.7
redis==5.0.1
pandas==2.1.3
numpy==1.26.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
asyncio==3.4.3
loguru==0.7.2
matplotlib==3.8.2
psutil==5.9.6
requests==2.31.0
scikit-learn==1.3.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
schedule==1.2.1
black==23.11.0
flake8==6.1.0 